init default sortInfo
